backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
waitress>=3.0,<4
pytest>=8.0,<9
pytest-django>=4.8,<5
